delete restriction
Test backend ready, doing front end

+++ src/main/java/se/hig/exte/service/AcademyService.java
@@ -2,14 +2,17 @@ package se.hig.exte.service;
 
 import java.util.List;
 
+import javax.security.sasl.AuthenticationException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import se.hig.exte.model.Academy;
 import se.hig.exte.repository.AcademyRepository;
+import se.hig.exte.restcontroller.SessionHandler;
 
 @Service
-public class AcademyService implements IService<Academy> {
+public class AcademyService {
 
 	private final AcademyRepository academyRepo;
 	
@@ -18,12 +21,12 @@ public class AcademyService implements IService<Academy> {
 		this.academyRepo = academyRepo;
 	}
 
-	@Override
+	//@Override
 	public Academy save(Academy academy) {
 		return academyRepo.save(academy);
 	}
 
-	@Override
+	//@Override
 	public Academy findById(int id) {
 		return academyRepo.findById(id);
 	}
@@ -32,8 +35,11 @@ public class AcademyService implements IService<Academy> {
 		return academyRepo.findAll();
 	}
 
-	@Override
-	public void deleteById(int id) {
-		academyRepo.deleteById(id);
+	public void deleteById(int id, String sessionId) throws AuthenticationException {
+		if(SessionHandler.isValidSuperSession(sessionId))
+			academyRepo.deleteById(id);
+		else 
+			throw new AuthenticationException();
 	}
+
 }
